Add edge case of bitslicing the whole int

+++ src/main/kotlin/simulator/impl/utils.kt
@@ -5,6 +5,6 @@ fun signExtend(v: Int, sz: Int): Int = v shl (32 - sz) shr (32 - sz)
 
 /** Returns x[end:start] = y */
 fun setBitslice(x: Int, y: Int, start: Int, end: Int): Int {
-    val mask: Int = ((1 shl end) - (1 shl start)).inv()
+    val mask: Int = ((1L shl end) - (1L shl start)).inv().toInt()
     return (mask and x) or (y shl start)
 }
